storage: consolidate page handling modules into `storage`
